splits tokeniser into its own module and changes method signatures in counter to accept a list
